MADS_NO_PYTHON and MADS_NO_PLOT

+++ examples/bayesian_sampling/runtests.jl
@@ -10,19 +10,25 @@ end
 md = Mads.loadmadsfile(workdir * "internal-linearmodel.mads")
 rootname = Mads.getmadsrootname(md)
 mcmcchain = Mads.bayessampling(md; nsteps=10, burnin=10, thinning=10)
-Mads.scatterplotsamples(md, mcmcchain.value', rootname * "-test-mcmcchain1.svg")
+if !haskey(ENV, "MADS_NO_PLOT")
+	Mads.scatterplotsamples(md, mcmcchain.value', rootname * "-test-mcmcchain1.svg")
+end
 mcmcchains = Mads.bayessampling(md, 2; nsteps=10, burnin=0)
-Mads.scatterplotsamples(md, vcat(map(chain->chain.value', mcmcchains)...), workdir * rootname * "-test-mcmcchain3.svg")
+if !haskey(ENV, "MADS_NO_PLOT")
+	Mads.scatterplotsamples(md, vcat(map(chain->chain.value', mcmcchains)...), workdir * rootname * "-test-mcmcchain3.svg")
+end
 
 md = Mads.loadmadsfile(workdir * "w01.mads")
 rootname = Mads.getmadsrootname(md)
 mcmcchain = Mads.bayessampling(md; nsteps=10, burnin=10, thinning=10)
-Mads.scatterplotsamples(md, mcmcchain.value', rootname * "-test-bayes-results.svg")
-mcmcvalues = Mads.paramarray2dict(md, mcmcchain.value') # convert the parameters in the chain to a parameter dictionary of arrays
-Mads.spaghettiplots(md, mcmcvalues, keyword="test")
-Mads.spaghettiplot(md, mcmcvalues, keyword="test")
-Mads.spaghettiplots(md, 3, keyword="test")
-Mads.spaghettiplot(md, 3, keyword="test")
-s = rootname
-run(`bash -c "rm -f $(s)*-test-*.svg"`)
+if !haskey(ENV, "MADS_NO_PLOT")
+	Mads.scatterplotsamples(md, mcmcchain.value', rootname * "-test-bayes-results.svg")
+	mcmcvalues = Mads.paramarray2dict(md, mcmcchain.value') # convert the parameters in the chain to a parameter dictionary of arrays
+	Mads.spaghettiplots(md, mcmcvalues, keyword="test")
+	Mads.spaghettiplot(md, mcmcvalues, keyword="test")
+	Mads.spaghettiplots(md, 3, keyword="test")
+	Mads.spaghettiplot(md, 3, keyword="test")
+	s = rootname
+	run(`bash -c "rm -f $(s)*-test-*.svg"`)
+end
 return
